Fix issue with secondary clients silently connecting
There is no easy way to limit a server to a single TCP connection
(the listen backlog is apparently not mandatory and does not help
under linux).

A second client is able to connect (but the connection is largely
dormant because it is not accepted). The client may even successfully
send data!.

We now let xvcSrv accept such connections and immediately close them
down thus giving the new clients a hint that something is wrong...